requirements.txt
fastapi
uvicorn
numpy
pandas
python-dotenv
langchain-core
langchain-openai
langchain-community
prometheus_fastapi_instrumentator
langchain_milvus
langgraph
transformers
FlagEmbedding
peft
ftfy
regex
tqdm
rocketmq-client-python
mysql-connector-python
redis
pika
requests
loguru



